cleanup(core): declutter lock file utils
