+++ packages/nx/src/lock-file/utils/lock-file-type.ts
